requirements.txt
alembic==1.6.5
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.2.0
blinker==1.4
click==7.1.2
colorama==0.4.4
Flask==1.1.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-restx==0.5.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
toml==0.10.2
Werkzeug==1.0.1

requests~=2.26.0
pandas~=1.3.1